fix: DailyStockChart tradingVolume 타입 수정

+++ BE/mmmm/src/main/java/com/beautifulyomin/mmmm/domain/stock/dto/response/StockResponse.java
@@ -21,5 +21,5 @@ public interface StockResponse {
 
     BigInteger getMarketCapitalization(); //시가총액
 
-    BigInteger getTradingVolume(); //누적 거래량
+    Long getTradingVolume(); //누적 거래량
 }
